Move mutex out of common (for now)
